requirements.txt
# Core dependencies
pandas>=2.0.0
langchain-ollama>=0.1.0
pydantic>=2.0.0
rich>=13.0.0
requests>=2.28.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional dependencies for enhanced features
python-dotenv>=1.0.0  # For environment variable management
pyyaml>=6.0.0  # For YAML configuration files